cli/requirements.txt
pyyaml==6.0.0
kubernetes==12.0.1
